# backend/internal/api/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: Parachute API
  description: |
    REST API for Parachute - A second brain app with Claude AI integration.

    Parachute provides a local-first, vault-based system for managing AI conversations,
    voice recordings, and knowledge across multiple spaces. Each space acts as a cognitive
    context with its own CLAUDE.md system prompt and SQLite knowledge database.

    ## Architecture
    - **Backend**: Go (Fiber v3) + SQLite
    - **Frontend**: Flutter + Riverpod
    - **AI Integration**: Claude via Agent Client Protocol (ACP)
    - **Data Storage**: Local vault (default: ~/Parachute/)

    ## Key Concepts
    - **Spaces**: Cognitive contexts with conversations and linked notes
    - **Captures**: Voice recordings with transcripts
    - **Conversations**: Chat sessions with Claude within a space
    - **Messages**: User and assistant messages in conversations
    - **Registry**: Global index of spaces and captures

    ## Authentication
    Currently uses ANTHROPIC_API_KEY environment variable or ~/.claude/.credentials.json

  version: 1.0.0
  contact:
    name: Parachute Support
    url: https://github.com/ParachuteLabs/parachute
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/spaces:
    get:
      summary: List all spaces
      description: Returns all spaces for the authenticated user
      tags:
        - Spaces
      responses:
        '200':
          description: List of spaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  spaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Space'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new space
      description: Creates a new space with optional icon and color
      tags:
        - Spaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceParams'
      responses:
        '201':
          description: Space created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}:
    get:
      summary: Get a space by ID
      description: Returns a single space
      tags:
        - Spaces
      parameters:
        - $ref: '#/components/parameters/SpaceID'
      responses:
        '200':
          description: Space details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update a space
      description: Updates space name, icon, or color
      tags:
        - Spaces
      parameters:
        - $ref: '#/components/parameters/SpaceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceParams'
      responses:
        '200':
          description: Space updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a space
      description: Deletes a space and its data
      tags:
        - Spaces
      parameters:
        - $ref: '#/components/parameters/SpaceID'
      responses:
        '204':
          description: Space deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}/notes:
    get:
      summary: Get notes linked to a space
      description: Returns all notes linked to this space with optional filtering
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "important,project-x"
        - name: start_date
          in: query
          description: Filter notes after this date (RFC3339)
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Filter notes before this date (RFC3339)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of notes to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of notes to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      $ref: '#/components/schemas/RelevantNote'
                  total:
                    type: integer
                    example: 42
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Link a note to a space
      description: Links an existing capture/note to this space with context and tags
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkNoteRequest'
      responses:
        '201':
          description: Note linked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "note linked successfully"
                  space_id:
                    type: string
                  capture_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}/notes/{capture_id}:
    put:
      summary: Update note context
      description: Updates the space-specific context and tags for a linked note
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
        - name: capture_id
          in: path
          required: true
          description: Capture ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteContextRequest'
      responses:
        '200':
          description: Note context updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "note context updated successfully"
                  space_id:
                    type: string
                  capture_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Unlink a note from a space
      description: Removes the link between a note and this space (does not delete the note file)
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
        - name: capture_id
          in: path
          required: true
          description: Capture ID
          schema:
            type: string
      responses:
        '200':
          description: Note unlinked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "note unlinked successfully"
                  space_id:
                    type: string
                  capture_id:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}/notes/{capture_id}/content:
    get:
      summary: Get note content
      description: Returns the full content of a linked note
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
        - name: capture_id
          in: path
          required: true
          description: Capture ID
          schema:
            type: string
      responses:
        '200':
          description: Note content
          content:
            application/json:
              schema:
                type: object
                properties:
                  capture_id:
                    type: string
                  note_path:
                    type: string
                    example: "captures/2025-10-26_00-00-17.md"
                  content:
                    type: string
                  space_context:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  linked_at:
                    type: string
                    format: date-time
                  last_referenced:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}/database/stats:
    get:
      summary: Get space database statistics
      description: Returns statistics about the space's SQLite knowledge database
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
      responses:
        '200':
          description: Database statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseStats'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/spaces/{id}/database/tables/{table_name}:
    get:
      summary: Query a space database table
      description: Returns raw data from a space database table
      tags:
        - Space Notes
      parameters:
        - $ref: '#/components/parameters/SpaceID'
        - name: table_name
          in: path
          required: true
          description: Table name
          schema:
            type: string
            example: "relevant_notes"
      responses:
        '200':
          description: Table data
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items:
                      type: string
                  rows:
                    type: array
                    items:
                      type: array
                      items: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/conversations:
    get:
      summary: List conversations
      description: Returns conversations for a specific space
      tags:
        - Conversations
      parameters:
        - name: space_id
          in: query
          required: true
          description: Space ID to filter conversations
          schema:
            type: string
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new conversation
      description: Creates a new conversation in a space
      tags:
        - Conversations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationParams'
      responses:
        '201':
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/conversations/{id}:
    get:
      summary: Get a conversation by ID
      description: Returns a single conversation
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update a conversation
      description: Updates conversation title
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Title"
      responses:
        '200':
          description: Conversation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a conversation
      description: Deletes a conversation and all its messages
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      responses:
        '204':
          description: Conversation deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/messages:
    get:
      summary: List messages
      description: Returns messages for a specific conversation
      tags:
        - Messages
      parameters:
        - name: conversation_id
          in: query
          required: true
          description: Conversation ID
          schema:
            type: string
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Send a message
      description: |
        Sends a user message to Claude and returns immediately.
        The assistant's response will be streamed via WebSocket.

        Connect to `/ws` and subscribe to the conversation to receive:
        - `message_chunk`: Text chunks of the assistant's response
        - `tool_call`: Tool call initiated by Claude
        - `tool_call_update`: Tool call status update
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: User message created successfully (assistant response will stream via WebSocket)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/captures/upload:
    post:
      summary: Upload a voice capture
      description: Uploads audio file with metadata
      tags:
        - Captures
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
                - timestamp
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (WAV format)
                timestamp:
                  type: string
                  format: date-time
                  description: Recording timestamp (RFC3339)
                duration:
                  type: number
                  format: float
                  description: Recording duration in seconds
                source:
                  type: string
                  description: Recording source
                  example: "omi-device"
                deviceId:
                  type: string
                  description: Device ID
                  example: "omi-123456"
      responses:
        '201':
          description: Capture uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  path:
                    type: string
                    example: "captures/2025-10-26_00-00-17.wav"
                  url:
                    type: string
                    example: "/api/captures/2025-10-26_00-00-17.wav"
                  createdAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/captures:
    get:
      summary: List captures
      description: Returns paginated list of captures
      tags:
        - Captures
      parameters:
        - name: limit
          in: query
          description: Maximum number of captures to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of captures to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of captures
          content:
            application/json:
              schema:
                type: object
                properties:
                  captures:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaptureMetadata'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  hasMore:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/captures/{filename}:
    get:
      summary: Download capture audio
      description: Downloads audio file
      tags:
        - Captures
      parameters:
        - name: filename
          in: path
          required: true
          description: Audio filename
          schema:
            type: string
            example: "2025-10-26_00-00-17.wav"
      responses:
        '200':
          description: Audio file
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a capture
      description: Deletes audio file and associated files (transcript, metadata)
      tags:
        - Captures
      parameters:
        - name: filename
          in: path
          required: true
          description: Audio filename
          schema:
            type: string
            example: "2025-10-26_00-00-17.wav"
      responses:
        '200':
          description: Capture deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/captures/{filename}/transcript:
    get:
      summary: Get capture transcript
      description: Returns transcript markdown
      tags:
        - Captures
      parameters:
        - name: filename
          in: path
          required: true
          description: Audio filename
          schema:
            type: string
            example: "2025-10-26_00-00-17.wav"
      responses:
        '200':
          description: Transcript markdown
          content:
            text/markdown:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/files/browse:
    get:
      summary: Browse files
      description: Lists files and directories in the vault
      tags:
        - Files
      parameters:
        - name: path
          in: query
          description: Relative path to browse (empty for vault root)
          schema:
            type: string
            example: "spaces/my-project"
      responses:
        '200':
          description: File listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    SpaceID:
      name: id
      in: path
      required: true
      description: Space ID
      schema:
        type: string
        example: "space-abc123"
  schemas:
    Space:
      type: object
      properties:
        id:
          type: string
          example: "space-abc123"
        user_id:
          type: string
          example: "default"
        name:
          type: string
          example: "My Project"
        path:
          type: string
          example: "/Users/john/Parachute/spaces/my-project"
        icon:
          type: string
          example: "🚀"
        color:
          type: string
          example: "#2E7D32"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...